Move organisms to pages and extract DashboardSelector
