+++ src/stories/DataExplorer.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Meta, Story } from '@storybook/react/types-6-0';
-import DataExplorer from '../components/organisms/DataExplorer';
+import DataExplorer from '../components/pages/DataExplorer';
 
 export default {
     title: 'DataExplorer',
